Missing features implemented Updated output formatting
