+++ tests/golem/test_client.py
@@ -3,21 +3,18 @@ import unittest
 import uuid
 
 from ethereum.utils import denoms
-from golem.task.taskmanager import TaskManager
-
-from golem.task.taskcomputer import TaskComputer
-from mock import Mock, MagicMock, patch
-
 from gnr.gnrapplicationlogic import GNRClientRemoteEventListener
 from golem.client import Client, GolemClientRemoteEventListener, ClientTaskComputerEventListener
 from golem.clientconfigdescriptor import ClientConfigDescriptor
 from golem.ethereum.paymentmonitor import IncomingPayment
 from golem.network.p2p.node import Node
 from golem.resource.dirmanager import DirManager
+from golem.task.taskcomputer import TaskComputer
+from golem.task.taskmanager import TaskManager
 from golem.task.taskserver import TaskServer
-from golem.task.taskstate import ComputerState
 from golem.tools.testdirfixture import TestDirFixture
 from golem.tools.testwithdatabase import TestWithDatabase
+from mock import Mock, MagicMock, patch
 
 
 class TestCreateClient(TestDirFixture):
@@ -225,107 +222,6 @@ class TestClientRPCMethods(TestWithDatabase):
         new_node_name = str(uuid.uuid4())
         c.update_setting('node_name', new_node_name)
         assert c.config_desc.node_name == new_node_name
-
-    @patch('golem.network.p2p.node.Node.collect_network_info')
-    def test_get_tasks(self, _):
-        c = self.__new_client()
-
-        c.task_server = TaskServer.__new__(TaskServer)
-        c.task_server.task_manager = TaskManager.__new__(TaskManager)
-        c.task_server.task_computer = TaskComputer.__new__(TaskComputer)
-        c.task_server.task_computer.dir_manager = DirManager(self.tempdir)
-        c.task_server.task_computer.current_computations = []
-
-        count = 3
-
-        tasks, tasks_states, task_id, subtask_id = self.__build_tasks(count)
-
-        c.task_server.task_manager.tasks = tasks
-        c.task_server.task_manager.tasks_states = tasks_states
-        c.task_server.task_manager.subtask2task_mapping = self.__build_subtask2task(tasks)
-
-        one_task = c.get_tasks(task_id)
-        assert one_task
-        assert isinstance(one_task, dict)
-        assert len(one_task)
-
-        all_tasks = c.get_tasks()
-
-        assert all_tasks
-        assert isinstance(all_tasks, list)
-        assert len(all_tasks) ==  count
-        assert all([isinstance(t, dict) for t in all_tasks])
-
-        one_subtask = c.get_subtask(subtask_id)
-
-        assert one_subtask
-        assert isinstance(one_subtask, dict)
-        assert len(one_subtask)
-
-        task_subtasks = c.get_subtasks(task_id)
-
-        assert task_subtasks
-        assert isinstance(task_subtasks, list)
-        assert all([isinstance(t, dict) for t in task_subtasks])
-
-    @classmethod
-    def __build_tasks(cls, n):
-
-        tasks = dict()
-        tasks_states = dict()
-        task_id = None
-        subtask_id = None
-
-        for i in xrange(0, n):
-
-            task = Mock()
-            task.header.task_id = str(uuid.uuid4())
-            task.get_total_tasks.return_value = i + 2
-            task.get_progress.return_value = i * 10
-
-            state = Mock()
-            state.status = 'waiting'
-            state.remaining_time = 100 - i
-
-            subtask_states, subtask_id = cls.__build_subtasks(n)
-
-            state.subtask_states = subtask_states
-            task.subtask_states = subtask_states
-
-            task_id = task.header.task_id
-
-            tasks[task.header.task_id] = task
-            tasks_states[task.header.task_id] = state
-
-        return tasks, tasks_states, task_id, subtask_id
-
-    @staticmethod
-    def __build_subtasks(n):
-
-        subtasks = dict()
-        subtask_id = None
-
-        for i in xrange(0, n):
-
-            subtask = Mock()
-            subtask.subtask_id = str(uuid.uuid4())
-            subtask.computer = ComputerState()
-            subtask.computer.node_name = 'node_{}'.format(i)
-            subtask.computer.node_id = 'deadbeef0{}'.format(i)
-            subtask_id = subtask.subtask_id
-
-            subtasks[subtask.subtask_id] = subtask
-
-        return subtasks, subtask_id
-
-    @staticmethod
-    def __build_subtask2task(tasks):
-        subtask2task = dict()
-        for k, t in tasks.items():
-            print k, t.subtask_states
-            for sk, st in t.subtask_states.items():
-                subtask2task[st.subtask_id] = t.header.task_id
-        return subtask2task
 
     def __new_client(self):
         client = Client(datadir=self.path,
